+++ src/packages/piral-blazor/src/create.ts
@@ -1,36 +1,47 @@
 import { Extend } from 'piral-core';
-import { initializeBlazor, eventNames } from './internal';
+import { initialize, addReference, activate, deactivate, attachEvents } from './internal';
 import { PiletBlazorApi } from './types';
 
 /**
  * Available configuration options for the Blazor plugin.
  */
-export interface BlazorConfig {}
+export interface BlazorConfig {
+  /**
+   * Determines if Blazor should only be loaded on demand.
+   * @default true
+   */
+  lazy?: boolean;
+}
 
 /**
  * Creates new Pilet API extensions for integration of Blazor.
  */
 export function createBlazorApi(config: BlazorConfig = {}): Extend<PiletBlazorApi> {
+  const { lazy = true } = config;
+  const bootConfig = require('../infra.codegen');
+  const boot = () => initialize(bootConfig);
+
   return context => {
     const root = document.body.appendChild(document.createElement('div'));
     root.style.display = 'none';
     root.id = 'blazor-root';
 
-    let loader: Promise<any>;
+    let loader = !lazy && boot();
 
-    context.converters.blazor = ({ moduleName, args }) => {
+    context.converters.blazor = ({ moduleName, args, dependency }) => {
       let id: string;
       let referenceId: string;
       let node: HTMLElement;
-      let renderHandler: (ev: CustomEvent) => void;
-      let navigateHandler: (ev: CustomEvent) => void;
+      let dispose = () => {};
       let state: 'fresh' | 'mounted' | 'removed';
 
       return {
         mount(el, data, ctx) {
           const props = { ...args, ...data };
-          loader
-            .then(() => DotNet.invokeMethodAsync<string>('NewBlazorApp', 'Activate', moduleName, props))
+
+          (loader || (loader = boot()))
+            .then(dependency)
+            .then(() => activate(moduleName, props))
             .then(refId => {
               if (state === 'fresh') {
                 id = `${moduleName}-${refId}`;
@@ -40,19 +51,19 @@ export function createBlazorApi(config: BlazorConfig = {}): Extend<PiletBlazorAp
               }
             })
             .catch(err => console.error(err));
-          renderHandler = ev => data.piral.renderHtmlExtension(ev.detail.target, ev.detail.props);
-          navigateHandler = ev => ctx.router.history.push(ev.detail.to);
-          el.addEventListener(eventNames.render, renderHandler, false);
-          el.addEventListener(eventNames.navigate, navigateHandler, false);
+          dispose = attachEvents(
+            el,
+            ev => data.piral.renderHtmlExtension(ev.detail.target, ev.detail.props),
+            ev => ctx.router.history.push(ev.detail.to),
+          );
           state = 'fresh';
         },
         unmount(el) {
-          el.removeEventListener(eventNames.render, renderHandler, false);
-          el.removeEventListener(eventNames.navigate, navigateHandler, false);
+          dispose();
 
           if (state === 'mounted') {
             root.querySelector(`#${id}`).appendChild(node);
-            DotNet.invokeMethodAsync('NewBlazorApp', 'Deactivate', moduleName, referenceId);
+            deactivate(moduleName, referenceId);
           }
 
           el.innerHTML = '';
@@ -61,18 +72,31 @@ export function createBlazorApi(config: BlazorConfig = {}): Extend<PiletBlazorAp
       };
     };
 
-    return {
-      setupBlazor(cfg) {
-        cfg.cacheBootResources = false;
-        loader = initializeBlazor(cfg);
-      },
-      fromBlazor(moduleName, args) {
-        return {
-          type: 'blazor',
-          moduleName,
-          args,
-        };
-      },
+    return () => {
+      let dependency: () => Promise<any>;
+
+      return {
+        defineBlazorReferences(references) {
+          const load = () =>
+            Promise.all(
+              references.map(reference =>
+                fetch(reference)
+                  .then(res => res.blob())
+                  .then(addReference),
+              ),
+            );
+          let result = !lazy && load();
+          dependency = () => result || (result = load());
+        },
+        fromBlazor(moduleName, args) {
+          return {
+            type: 'blazor',
+            dependency,
+            moduleName,
+            args,
+          };
+        },
+      };
     };
   };
 }