+++ src/packages/piral-blazor/src/internal/index.ts
@@ -1,6 +1,7 @@
 import './globals';
 import { setPlatform } from './Environment';
 import { Pointer } from './Platform/Platform';
+import { navigationManager } from './Services/NavigationManager';
 import { BootJsonData, BootConfigResult } from './Platform/BootConfig';
 import { monoPlatform } from './Platform/Mono/MonoPlatform';
 import { WebAssemblyResourceLoader } from './Platform/WebAssemblyResourceLoader';
@@ -17,7 +18,55 @@ export const eventNames = {
   navigate: 'navigate-blazor',
 };
 
-export async function initializeBlazor(data: BootJsonData) {
+const coreLib = 'Piral.Blazor.Core';
+
+function isRooted(target: HTMLElement) {
+  let parent = target.parentElement;
+
+  while (parent) {
+    if (parent.id === 'blazor-root') {
+      return true;
+    }
+
+    parent = parent.parentElement;
+  }
+
+  return false;
+}
+
+export function activate(moduleName: string, props: any) {
+  return DotNet.invokeMethodAsync<string>(coreLib, 'Activate', moduleName, props);
+}
+
+export function deactivate(moduleName: string, referenceId: string) {
+  return DotNet.invokeMethodAsync<string>(coreLib, 'Deactivate', moduleName, referenceId);
+}
+
+export function addReference(blob: Blob) {
+  return new Promise(resolve => {
+    var reader = new FileReader();
+    reader.onload = () => {
+      const data = reader.result.toString().replace(/^data:.+;base64,/, '');
+      DotNet.invokeMethodAsync(coreLib, 'LoadComponentsFromLibrary', data).then(resolve);
+    };
+    reader.readAsDataURL(blob);
+  });
+}
+
+export function attachEvents(
+  host: HTMLElement,
+  render: (ev: CustomEvent) => void,
+  navigate: (ev: CustomEvent) => void,
+) {
+  host.addEventListener(eventNames.render, render, false);
+  host.addEventListener(eventNames.navigate, navigate, false);
+  return () => {
+    host.removeEventListener(eventNames.render, render, false);
+    host.removeEventListener(eventNames.navigate, navigate, false);
+  };
+}
+
+export async function initialize(data: BootJsonData) {
   data.cacheBootResources = false;
 
   setEventDispatcher((eventDescriptor, eventArgs) =>
@@ -31,19 +80,6 @@ export async function initializeBlazor(data: BootJsonData) {
 
   // Configure environment for execution under Mono WebAssembly with shared-memory rendering
   const platform = setPlatform(monoPlatform);
-  const isRooted = (target: HTMLElement) => {
-    let parent = target.parentElement;
-
-    while (parent) {
-      if (parent.id === 'blazor-root') {
-        return true;
-      }
-
-      parent = parent.parentElement;
-    }
-
-    return false;
-  };
 
   Object.assign(window['Blazor'], {
     platform,
@@ -80,6 +116,7 @@ export async function initializeBlazor(data: BootJsonData) {
   });
 
   Object.assign(window['Blazor']._internal, {
+    navigationManager,
     attachRootComponentToElement,
     renderBatch(browserRendererId: number, batchAddress: Pointer) {
       renderBatch(browserRendererId, new SharedMemoryRenderBatch(batchAddress));