+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.11.2 (tbd)
 
+* Added new optional plugin `piral-blazor` for Blazor (#112)
 * Improved scaffolding of files for pilets (#189)
 * Fixed verbose logging on Windows (#192)
 * Fixed the returned type of the `piral-feeds` HOC (#193)